% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init}
\alias{init}
\title{Initialize gitlatex}
\usage{
init(
  public = "./gitlatex.json",
  private = "./gitlatex.private.json",
  assets = "./assets"
)
}
\arguments{
\item{public}{Path to the public config file (default: \code{./gitlatex.json})}

\item{private}{Path to the private config file (default: \code{./gitlatex.private.json})}

\item{assets}{Path to the assets directory (default: \code{./assets})}
}
\description{
Initializes gitlatex in a new or existing project, by creating the relevant config files and directories.
}
