% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_to_assets.R
\name{link_to_assets}
\alias{link_to_assets}
\title{Create symlinks to the assets directory}
\usage{
link_to_assets(config, projects = NULL)
}
\arguments{
\item{config}{The config object, should be read using \code{\link[=read_config]{read_config()}}}

\item{projects}{A character vector of project names. If \code{NULL}, all projects will be linked.}
}
\description{
Creates symlinks to the assets directory in each project local directory
}
