% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/push.R
\name{push}
\alias{push}
\title{Push changes to remote project directories}
\usage{
push(
  config,
  projects = NULL,
  main = TRUE,
  latest = TRUE,
  in_local = TRUE,
  more_recent = TRUE,
  git = TRUE
)
}
\arguments{
\item{config}{The config object, should be read using \code{\link[=read_config]{read_config()}}}

\item{projects}{A character vector of project names. If \code{NULL}, all projects will be pushed.}

\item{main}{Should we test that the repo is on the main branch before pushing? (default: \code{TRUE})}

\item{latest}{Should we test that the repo is up to date with master before pushing? (default: \code{TRUE})}

\item{in_local}{Should we test that all files in remote are also in local? (default: \code{TRUE}; this ignores LaTeX build files)}

\item{more_recent}{Should we test that all files in local are more recent than in remote? (default: \code{TRUE})}

\item{git}{Should we run the Git tests at all? (default: \code{TRUE})}
}
\description{
Pushes changes to remote project directories.
}
\details{
The logical parameters issue warnings when turned off.
If turned on, they issue errors in non-interactive sessions and ask for user confirmation in interactive sessions.
All tests are run, even when turned off. The \code{git} parameter allows to skip the Git tests altogether.
This is useful in cases where the local directory is not a Git repo.
}
