% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_config.R
\name{read_config}
\alias{read_config}
\title{Read gitlatex config files}
\usage{
read_config(public, private)
}
\arguments{
\item{public}{Path to the public config file (default: \code{./gitlatex.json})}

\item{private}{Path to the private config file (default: \code{./gitlatex.private.json})}
}
\value{
A config object to be used by other functions
}
\description{
Reads the public and private config files for gitlatex. Those are created using \code{\link[=init]{init()}}.
}
