% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull.R
\name{pull}
\alias{pull}
\title{Pull changes from remote project directories}
\usage{
pull(config, projects = NULL, uncommitted = TRUE)
}
\arguments{
\item{config}{The config object, should be read using \code{\link[=read_config]{read_config()}}}

\item{projects}{A character vector of project names. If \code{NULL}, all projects will be pulled.}

\item{uncommitted}{Should we test that the repo has no uncommitted changes before pulling? (default: \code{TRUE})}
}
\description{
Pulls changes from remote project directories.
}
\details{
The logical parameter issues warnings when turned off.
If turned on, they issue errors in non-interactive sessions and ask for user confirmation in interactive sessions.
}
